requirements.txt
Django>=3.2.5,<3.3.0
djangorestframework>=3.13.1,<3.14.9
djangorestframework-jwt>=1.11.0,<1.12.0
django-cors-headers>=3.11.0,<3.12.0
django-rest-auth>=0.9.5,<0.10.0
django-allauth>=0.49.0,<0.50.0

psycopg2>=2.9.3,<2.10.0
python-dotenv>=0.19.2,<0.20.0
drf-yasg>=1.20.0,<1.21.0
djoser>=2.1.0,<2.2.0